Move theme output generation to separate composable
