+++ packages/themes/src/index.ts
@@ -1,3 +1,4 @@
-export { ThemeSchema, Definitions, TypeId, type Theme } from './schema.js';
+export { Definitions, ThemeSchema, TypeId, type Theme } from './schema.js';
 export { useThemeStore } from './store.js';
 export { default as ThemeProvider } from './theme-provider.vue';
+export { useTheme } from './use-theme.js';
